Query subgraph directly from the client as API key doesn't need to be hidden because the requests can be restricted to specific domain
