+++ packages/lib/apollo/client.ts
@@ -2,7 +2,7 @@ import { ApolloClient, InMemoryCache } from "@apollo/client";
 
 const client = new ApolloClient({
   cache: new InMemoryCache(),
-  uri: `${process.env.SUBGRAPH_ENDPOINT}`,
+  uri: `${process.env.NEXT_PUBLIC_SUBGRAPH_ENDPOINT}`,
 });
 
 export default client;
