+++ packages/ui/hooks/useAuction.ts
@@ -1,5 +1,7 @@
 import useSWR, { SWRResponse } from "swr";
 import { zeroAddress } from "viem";
+import client from "lib/apollo/client";
+import { GET_SALE_QUERY } from "lib/apollo/query";
 
 const useAuction = (
   id: `0x${string}`,
@@ -9,12 +11,17 @@ const useAuction = (
     address,
   }).toString();
 
-  const fetcher = (url: string): Promise<any | undefined> =>
-    fetch(url)
-      .then((res) => res.json())
-      .then((result) => {
-        return { auction: result.data.auction };
-      });
+  const fetcher = async (key: string): Promise<any | undefined> => {
+    const result = await client.query({
+      query: GET_SALE_QUERY,
+      variables: {
+        id: id as string,
+        address: (address as `0x${string}`).toLowerCase(),
+      },
+    });
+    return { auction: result.data.auction };
+  };
+
   return useSWR<any | undefined, Error>(`/api/auctions/${id}?${params}`, fetcher, {
     errorRetryCount: 2,
   });
