+++ packages/ui/hooks/useAuctions.ts
@@ -1,7 +1,17 @@
+import { DocumentNode } from "@apollo/client";
 import { KeyedMutator, SWRConfiguration } from "swr";
 import useSWRInfinite from "swr/infinite";
-import { QueryType } from "lib/apollo/query";
+import {
+  QueryType,
+  LIST_ACTIVE_AND_UPCOMING_SALE_QUERY,
+  LIST_ACTIVE_SALE_QUERY,
+  LIST_UPCOMING_SALE_QUERY,
+  LIST_CLOSED_SALE_QUERY,
+  LIST_MY_SALE_QUERY,
+  LIST_PARTICIPATED_SALE_QUERY,
+} from "lib/apollo/query";
 import { AuctionProps } from "lib/types/Auction";
+import client from "lib/apollo/client";
 import { zeroAddress } from "viem";
 import { useCallback } from "react";
 
@@ -30,28 +40,47 @@ export const useSWRAuctions = (
   config: AuctionsParams & SWRConfiguration,
   queryType: QueryType = QueryType.ACTIVE_AND_UPCOMING,
 ): SWRAuctionStore => {
+  const getQuery = (queryType: QueryType): DocumentNode => {
+    switch (queryType) {
+      case QueryType.ACTIVE_AND_UPCOMING:
+        return LIST_ACTIVE_AND_UPCOMING_SALE_QUERY;
+      case QueryType.ACTIVE:
+        return LIST_ACTIVE_SALE_QUERY;
+      case QueryType.UPCOMING:
+        return LIST_UPCOMING_SALE_QUERY;
+      case QueryType.CLOSED:
+        return LIST_CLOSED_SALE_QUERY;
+      case QueryType.MY_SALE_QUERY:
+        return LIST_MY_SALE_QUERY;
+      case QueryType.PARTICIPATED_SALE_QUERY:
+        return LIST_PARTICIPATED_SALE_QUERY;
+      default:
+        return LIST_ACTIVE_AND_UPCOMING_SALE_QUERY;
+    }
+  };
   const getKey = (pageIndex: number, previousPageData: AuctionProps[]) => {
     if (previousPageData && !previousPageData.length) return null;
-    const skip = pageIndex * LIMIT;
+    let skip = pageIndex * LIMIT;
+    skip = config.skip ? config.skip + skip : skip;
+    const query = getQuery(queryType);
+    const first = config.first ? config.first : LIMIT;
+    const id = config.id ? config.id : zeroAddress;
+    const now = NOW;
 
-    const params = new URLSearchParams({
-      queryTypeIndex: queryType.toString(),
-      skip: (config.skip ? config.skip + skip : skip).toString(),
-      first: (config.first ? config.first : LIMIT).toString(),
-      id: (config.id ? config.id : zeroAddress).toString(),
-      now: NOW.toString(),
-    }).toString();
-
-    return params;
+    return { query, variables: { skip, first, now, id } };
   };
 
-  const fetcher = async (params: string) => {
+  const fetcher = async (
+    params: {
+      query: DocumentNode;
+      variables: { skip: number; first: number; now: number; id: string };
+    } | null,
+  ) => {
     let auctions: AuctionProps[] = [];
+    if (params === null) return auctions;
     try {
-      const result = await fetch(`/api/auctions/?${params}`);
-      if (!result.ok) throw new Error();
-      const json = await result.json();
-      auctions = json.data.auctions;
+      const result = await client.query(params);
+      auctions = result.data.auctions;
     } catch (e) {
       console.error(e);
     }
