+++ packages/ui/hooks/useTemplates.ts
@@ -1,12 +1,15 @@
 import useSWR, { SWRResponse } from "swr";
+import { LIST_TEMPLATE_QUERY } from "lib/apollo/query";
+import client from "lib/apollo/client";
 
 const useTemplates = (): SWRResponse<any | undefined, Error> => {
-  const fetcher = (url: string): Promise<any | undefined> =>
-    fetch(url)
-      .then((res) => res.json())
-      .then((result) => {
-        return { templates: result.data.templates };
-      });
+  const fetcher = async (key: string): Promise<any | undefined> => {
+    const result = await client.query({
+      query: LIST_TEMPLATE_QUERY,
+    });
+    return { templates: result.data.templates };
+  };
+
   return useSWR<any | undefined, Error>(`/api/templates`, fetcher, {
     errorRetryCount: 2,
   });
