+++ turbo.json
@@ -23,7 +23,7 @@
         "AWS_DYNAMO_TABLE_NAME",
         "NODE_ENV",
         "NEXT_PUBLIC_VERCEL_GIT_COMMIT_SHA",
-        "SUBGRAPH_ENDPOINT"
+        "NEXT_PUBLIC_SUBGRAPH_ENDPOINT"
       ]
     },
     "lint": {},
